% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractForest.R
\name{extractForest}
\alias{extractForest}
\title{A function to extract forest areas from osm}
\usage{
extractForest(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "landuse" and the value "forest".
}
\description{
A function to extract forest areas from osm
}
\examples{
extractForest(aoi)
}
\keyword{internal}
