% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGarden.R
\name{extractGarden}
\alias{extractGarden}
\title{A function to extract public garden areas from osm}
\usage{
extractGarden(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "leisure" and the value "garden".
}
\description{
A function to extract public garden areas from osm
}
\examples{
extractGarden(aoi)
}
