% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGrassland.R
\name{extractGrassland}
\alias{extractGrassland}
\title{A function to extract grassland areas from OSM}
\usage{
extractGrassland(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "nature" and the values "grassland", "scrub" and "shrubbery".
}
\description{
A function to extract grassland areas from OSM
}
\examples{
extractGrassland(aoi)
}
