% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMeadow.R
\name{extractMeadow}
\alias{extractMeadow}
\title{A function to extract meadow areas from OSM}
\usage{
extractMeadow(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "landuse" and the value "meadow".
}
\description{
A function to extract meadow areas from OSM
}
\examples{
extractMeadow(aoi)
}
