% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPark.R
\name{extractPark}
\alias{extractPark}
\title{A function to extract park areas from OSM}
\usage{
extractPark(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "leisure" and the value "park".
}
\description{
A function to extract park areas from OSM
}
\examples{
extractPark(aoi)
}
