% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRiver.R
\name{extractRiver}
\alias{extractRiver}
\title{A function to extract rivers from OSM}
\usage{
extractRiver(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "waterway" and the value "river".
}
\description{
A function to extract rivers from OSM
}
\examples{
extractRiver(aoi)
}
