% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSingleTree.R
\name{extractSingleTree}
\alias{extractSingleTree}
\title{A function to extract single trees from OSM}
\usage{
extractSingleTree(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "natural" and the value "tree".
}
\description{
A function to extract single trees from OSM
}
\examples{
extractSingleTree(aoi)
}
\keyword{internal}
