% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSpecialForest.R
\name{extractSpecialForest}
\alias{extractSpecialForest}
\title{A function to extract forest specified as mixed, coniferous or deciduous from OSM}
\usage{
extractSpecialForest(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "leaf_type" and the values "mixed", "coniferous" and "deciduous".
}
\description{
A function to extract forest specified as mixed, coniferous or deciduous from OSM
}
\examples{
extractSpecialForest(aoi)
}
