% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractStreets.R
\name{extractStreets}
\alias{extractStreets}
\title{A function to extract streets from OSM}
\usage{
extractStreets(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "highway" and the values:
    "motorway", "trunk", "primary", "secondary",
    "tertiary", "unclassified", "residential",
    "motorway_link", "trunk_link", "primary_link",
    "secondary_link", "tertiary_link",
    "living_street", "service", "pedestrian",
    "track", "bus_guideway", "escape", "raceway",
    "road", "busway",
    "footway", "bridleway", "steps", "cooridor",
    "path", "via_ferrata",
    "ladder", "mini_roundabout", "motorway_junction",
    "turning_circle".
}
\description{
A function to extract streets from OSM
}
\examples{
extractStreets(aoi)
}
\keyword{internal}
