% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWine.R
\name{extractWine}
\alias{extractWine}
\title{A function to extract wine yards from OSM}
\usage{
extractWine(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "landuse" and the value "vineyard".
}
\description{
A function to extract wine yards from OSM
}
\examples{
extractWine(aoi)
}
\keyword{internal}
