% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWood.R
\name{extractWood}
\alias{extractWood}
\title{A function to extract wooda areas from OSM}
\usage{
extractWood(aoi)
}
\arguments{
\item{aoi}{An AOI feature created by spatialExtent() or another vector of a bounding box (crs 4326).}
}
\value{
A simple feature collection of OSM-features with the key "natural" and the value "wood".
}
\description{
A function to extract wooda areas from OSM
}
\examples{
extractWood(aoi)
}
\keyword{internal}
