% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalPlot.R
\name{finalPlot}
\alias{finalPlot}
\title{A function to plot the choosen OSM features}
\usage{
finalPlot(
  streets,
  river,
  forest,
  specialForest,
  wood,
  singleTree,
  treeLine,
  grassland,
  meadow,
  park,
  garden,
  natureReserve,
  pov,
  xlim,
  ylim,
  closestNature,
  route
)
}
\arguments{
\item{streets}{A simple feature collection of OSM-features.}

\item{river}{A simple feature collection of OSM-features.}

\item{forest}{A simple feature collection of OSM-features.}

\item{specialForest}{A simple feature collection of OSM-features.}

\item{wood}{A simple feature collection of OSM-features.}

\item{singleTree}{A simple feature collection of OSM-features.}

\item{treeLine}{A simple feature collection of OSM-features.}

\item{grassland}{A simple feature collection of OSM-features.}

\item{meadow}{A simple feature collection of OSM-features.}

\item{park}{A simple feature collection of OSM-features.}

\item{garden}{A simple feature collection of OSM-features.}

\item{natureReserve}{A simple feature collection of OSM-features.}

\item{pov}{A simple feature point.}

\item{xlim}{A vector.}

\item{ylim}{A vector.}

\item{closestNature}{A simple feature point.}

\item{route}{A simple feature line.}
}
\value{
The final plot of the simple feature collections, the users chosen position (POV),
    the closest point of public available nature and the route to get there.
}
\description{
A function to plot the choosen OSM features
}
\examples{
finalPlot(streets, river, forest,
          specialForest, wood, singleTree,
          treeLine, grassland,
          meadow, park, garden,
          natureReserve, pov, xlim, ylim, closestNature, route)
}
