% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNature.R
\name{findNature}
\alias{findNature}
\title{A function to find the closest point of public nature}
\usage{
findNature(
  streets,
  river,
  forest,
  specialForest,
  wood,
  singleTree,
  treeLine,
  grassland,
  meadow,
  park,
  garden,
  natureReserve,
  pov
)
}
\arguments{
\item{streets}{A simple feature collection of OSM-features.}

\item{river}{A simple feature collection of OSM-features.}

\item{forest}{A simple feature collection of OSM-features.}

\item{specialForest}{A simple feature collection of OSM-features.}

\item{wood}{A simple feature collection of OSM-features.}

\item{singleTree}{A simple feature collection of OSM-features.}

\item{treeLine}{A simple feature collection of OSM-features.}

\item{grassland}{A simple feature collection of OSM-features.}

\item{meadow}{A simple feature collection of OSM-features.}

\item{park}{A simple feature collection of OSM-features.}

\item{garden}{A simple feature collection of OSM-features.}

\item{natureReserve}{A simple feature collection of OSM-features.}

\item{pov}{A simple feature of type point.}
}
\value{
A sf object of the closest available public nature
}
\description{
A function to find the closest point of public nature
}
\examples{
findNature(streets, river, forest,
          specialForest, wood, singleTree,
          treeLine, grassland,
          meadow, park, garden,
          natureReserve, pov)
}
