% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POV.R
\name{getPOV}
\alias{getPOV}
\title{A simple function to get the position as sf object}
\usage{
getPOV(xcoord, ycoord)
}
\arguments{
\item{xcoord}{A x coordinate as decimal degree.}

\item{ycoord}{A y coordinate as decimal degree.}
}
\value{
The position determined by the two coordinates as
    sf object in the coordinate system 4326.
}
\description{
A simple function to get the position as sf object
}
\examples{
getPOV(8.671087, 52.113866)
}
\keyword{internal}
