% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereRthetrees.R
\name{whereRthetrees}
\alias{whereRthetrees}
\title{A function to find the closest public nature area around you}
\usage{
whereRthetrees(x, y, buffer)
}
\arguments{
\item{x}{A x coordinate as decimal degree.}

\item{y}{A y coordinate as decimal degree.}

\item{buffer}{The buffer radius around the position [m]. If too big,
an error will occur because this package is using the demo-server of osrm
which restricts the query size to 10000.}
}
\value{
A map of the nature areas around the given position, with a
    route pointing to the closest one.
}
\description{
A function to find the closest public nature area around you
}
\examples{
whereRtheTrees(8.671087, 52.113866, 300)
}
