% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyLimits.R
\name{xyLimits}
\alias{xyLimits}
\title{A function for calculating the x and y limits of an AOI}
\usage{
xyLimits(xcoord, ycoord, buffer)
}
\arguments{
\item{xcoord}{A x coordinate as decimal degree.}

\item{ycoord}{A y coordinate as decimal degree.}

\item{buffer}{The buffer radius around the position.}
}
\value{
A vector containing the x and y minima and maxima.
}
\description{
A function for calculating the x and y limits of an AOI
}
\examples{
xyLimits(8.671087, 52.113866, 300)
}
