% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereRthetrees.R
\name{whereRthetrees}
\alias{whereRthetrees}
\title{A function to find the closest public nature area around you}
\usage{
whereRthetrees(x, y, buffer)
}
\arguments{
\item{x}{A x coordinate as decimal degree.}

\item{y}{A y coordinate as decimal degree.}

\item{buffer}{The buffer radius around the position [m]. If too big,
an error will occur because this package is using the demo-server of osrm
which restricts the query size to 10000. Therfore, a buffer of 200-300m is
often sufficient.}
}
\value{
A map of the nature areas around the given position, with a
    route pointing to the closest one.
}
\description{
Being outside in nature is proven to be good for body and mind.
It can be calming, relaxing and grounding. It can make one happy and more level-headed.
So take a break, look for the closest public nature spot and go hug a tree!
}
\examples{
whereRtheTrees(8.671087, 52.113866, 300)
}
